requirements.txt
Django>=3.2.0,<3.9.0
djangorestframework>=3.12.4,<3.15.0
psycopg2>=2.8,<2.9
Pillow>=8.3.2,<8.5.0
flake8>=3.9.0,<3.10.0
autopep8>=1.5.6,<1.6.0
uWSGI>=2.0.20,<2.2